% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltrccf.R
\name{ltrccf}
\alias{ltrccf}
\title{Fit a LTRC conditional inference forest}
\usage{
ltrccf(
  formula,
  data,
  id,
  mtry = NULL,
  ntree = 100L,
  bootstrap = c("by.sub", "by.root", "by.user", "none"),
  perturb = list(replace = FALSE, fraction = 0.632),
  samp = NULL,
  trace = TRUE,
  stepFactor = 2,
  na.action = na.pass,
  applyfun = NULL,
  cores = NULL,
  control = partykit::ctree_control(teststat = "quad", testtype = "Univ", minsplit =
    max(ceiling(sqrt(nrow(data))), 20), minbucket = max(ceiling(sqrt(nrow(data))), 7),
    minprob = 0.01, mincriterion = 0, saveinfo = FALSE)
)
}
\arguments{
\item{formula}{a formula object, with the response being a \code{\link[survival]{Surv}}
object, with form


\code{Surv(tleft, tright, event)}.}

\item{data}{a data frame containing \code{n} rows of
left-truncated right-censored observations.
For time-varying data, this should be
a data frame containing pseudo-subject observations based on the Andersen-Gill
reformulation.}

\item{id}{variable name of subject identifiers. If this is present, it will be
searched for in the \code{data} data frame. Each group of rows in \code{data}
with the same subject \code{id} represents the covariate path through time of
a single subject. If not specified, the algorithm then assumes \code{data}
contains left-truncated and right-censored survival data with time-invariant
covariates.}

\item{mtry}{number of input variables randomly sampled as candidates at each node for
random forest like algorithms. The default \code{mtry} is tuned by \code{\link{tune.ltrccf}}.}

\item{ntree}{an integer, the number of the trees to grow for the forest.
\code{ntree = 100L} is set by default.}

\item{bootstrap}{bootstrap protocol.
(1) If \code{id} is present,
the choices are: \code{"by.sub"} (by default) which bootstraps subjects,
\code{"by.root"} which bootstraps pseudo-subjects.
Both can be with or without replacement (by default sampling is without
replacement; see the option \code{perturb} below);
(2) If \code{id} is not specified, it
bootstraps the \code{data} by sampling with or without replacement.
Regardless of the presence of \code{id}, if \code{"none"} is chosen, the
\code{data} is not bootstrapped at all. If \code{"by.user"} is choosen,
the bootstrap specified by \code{samp} is used.}

\item{perturb}{a list with arguments \code{replace} and \code{fraction} determining which
type of resampling, with \code{replace = TRUE} referring to the \emph{n}-out-of-\emph{n}
bootstrap and \code{replace = FALSE} referring to sample splitting. \code{fraction} is
the proportion of observations to draw without replacement.}

\item{samp}{Bootstrap specification when \code{bootstype = "by.user"}.
Array of dim \code{n x ntree} specifying how many times each record appears
inbag in the bootstrap for each tree.}

\item{trace}{whether to print the progress of the search of the optimal value of
\code{mtry}, when \code{mtry} is not specified (see \code{\link{tune.ltrccf}}).
\code{trace = TRUE} is set by default.}

\item{stepFactor}{at each iteration, \code{mtry} is inflated (or deflated)
by this value, used when \code{mtry} is not specified (see \code{\link{tune.ltrccf}}).
The default value is \code{2}.}

\item{na.action}{a function which indicates what should happen when the data contain
missing values.}

\item{applyfun}{an optional \code{lapply}-style function with arguments
\code{function(X, FUN, ...)}.
It is used for computing the variable selection criterion. The default is to use the
basic \code{lapply} function unless the \code{cores} argument is specified (see below).
See \code{\link[partykit]{ctree_control}}.}

\item{cores}{numeric. If set to an integer the \code{applyfun} is set to
\code{\link[parallel]{mclapply}} with the desired number of cores.
See \code{\link[partykit]{ctree_control}}.}

\item{control}{a list of control parameters, see \code{\link[partykit]{ctree_control}}.
\code{control} parameters \code{minsplit}, \code{minbucket} have been adjusted from the
\code{\link[partykit]{cforest}} defaults. Other default values correspond to those of the
default values used by \code{\link[partykit]{ctree_control}}.}
}
\value{
An object belongs to the class \code{ltrccf}, as a subclass of
\code{\link[partykit]{cforest}}.
}
\description{
An implementation of the random forest and bagging ensemble algorithms utilizing
LTRC conditional inference trees \code{\link{LTRCIT}} as base learners for
left-truncated right-censored survival data with time-invariant covariates.
It also allows for (left-truncated) right-censored survival data with
time-varying covariates.
}
\details{
This function extends the conditional inference survival forest algorithm in
\code{\link[partykit]{cforest}} to fit left-truncated and right-censored data,
which allow for time-varying covariates. The traditional survival forests in
\code{\link[partykit]{cforest}} only applies for right-censored data with
time-invariant invariant covariates.
}
\examples{
#### Example with time-varying data pbcsample
Formula = Surv(Start, Stop, Event) ~ age + alk.phos + ast + chol + edema
## Fit an LTRC conditional inference forest, with mtry specified
LTRCCFobj = ltrccf(formula = Formula, data = pbcsample, id = ID, mtry = 3, ntree = 50L)

## Fit an LTRC conditional inference forest, with mtry tuned with stepFactor=3.
LTRCCFobj = ltrccf(formula = Formula, data = pbcsample, id = ID, ntree = 50L, stepFactor = 3)
}
\references{
Andersen, P. and Gill, R. (1982). Cox's regression model for counting
processes, a large sample study. \emph{Annals of Statistics}, \strong{10}, 1100-1120.
}
\seealso{
\code{\link{predict.ltrccf}} for prediction and \code{\link{tune.ltrccf}}
for \code{mtry} tuning.
}
\keyword{Ensemble}
\keyword{conditional}
\keyword{covariate}
\keyword{data}
\keyword{data,}
\keyword{forest,}
\keyword{inference}
\keyword{left-truncated}
\keyword{method,}
\keyword{right-censored}
\keyword{time-varying}
